+++ lib/qt4/NatureListComboBox.cpp
@@ -21,7 +21,7 @@ namespace pkmn
             std::vector<pkstring> natures_vec;
             get_nature_list(natures_vec);
 
-            for(int i = 0; i < natures_vec.size(); i++)
+            for(size_t i = 0; i < natures_vec.size(); i++)
             {
                 addItem(tr(natures_vec[i].const_char()), QVariant(i));
             }
